Gapfill using georegional means

Gapfill using georegional means, providing the finest possible resolution from 3 hierarchies (r2 > r1 > r0) derived from United Nations geoscheme.

gapfill_georegions(data, georegions, fld_id = intersect(names(data),
  names(georegions)), fld_year = ifelse("year" %in% names(data), "year",
  NA), fld_value = setdiff(names(data), c(fld_id, fld_weight, "year")),
  georegion_labels = NULL, fld_weight = NULL, rgn_weights = NULL,
  ratio_weights = FALSE, gapfill_scoring_weights = c(r0 = 1, r1 = 0.8, r2 =
  0.5, v = 0), r0_to_NA = TRUE, attributes_csv = NULL)

Arguments

data
data.frame to gapfill having at least fields: fld_id and fld_value, and optionally fld_weight

georegions
data.frame having at least fields: fld_id and r0, r1, and r2 with georegion id values

fld_id
common spatial id field (eg region_id or country_key) between data and georegions

fld_year
optional year field in data

fld_value
value to gapfill in data

georegion_labels
with same dimensions as georegions having fields: r0_label, r1_label, r2_label and v_label

fld_weight
optional weighting field in data

rgn_weights
data frame of weights, expecting rgn_id in first column and weight in second

ratio_weights
if TRUE, multiply the gapfilled value by the ratio of the region's weight to the regional average weight. Defaults to FALSE. IMPORTANT to set to TRUE if dealing with values that SUM!

gapfill_scoring_weights
used to determine gapfilling scoreset. should range 0 to 1. defaults to c('r0'=1, 'r1'=0.8, 'r2'=0.5, 'v'=0)

r0_to_NA
assign value of NA if only georegional average availabe at the global level (r0). defaults to True.

attributes_csv
optional path and filename to save attribute table. defaults to NULL

Value

Returns a data.frame of having all the fld_id from georegions filled in the following columns:

fld_id - spatial id (eg region_id or country_key).

fld_value - the gapfilled value (eg score).

The returned data.frame also has an attribute "gapfill_georegions" which shows the calculated georegional means and which levels were chosen:

r0 - georegional id for level 0, ie global.

r1 - georegional id for level 1.

r2 - georegional id for level 2, the finest resolution of georegions.

id - spatial id (eg region_id or country_key).

w - weight used to apply weighted.mean. Defaults to 1 if not supplied as fld_weight parameter.

v - original fld_value in data

r2_v - weighted.mean for level 2

r1_v - weighted.mean for level 1

r0_v - weighted.mean for level 0 (global)

r2_n - count of regions available for level 2

r1_n - count of regions available for level 1

r0_n - count of regions available for level 0

r2_n_notna - count of region values that are not NA for level 2

r1_n_notna - count of region values that are not NA for level 1

r0_n_notna - count of region values that are not NA for level 0

z_level - finest level available

z_ids - ids for regions that are not NA which contributed to the score

z_n - count of input values for finest level available

z_n_pct - percent of region values that are not NA over all possible [0 to 1]

z_g_score - gapfilling score (see details)

z - weighted.mean for finest level available

Details

Gapfill using georegional means, providing the finest possible resolution from 3 hierarchies (r2 > r1 > r0).

The gapfill score (z_g_score) in the attribute table is formulated such that the higher the score, the more gapfilling performed. The maximal gapfill score is based on gapfilling at the global level (r0=1) and least if no gapfilling performed (ie z = v). But then some regional averages are applied with only a few regional values while others might have all but the gapfilled region available. To account for this aspect, the difference between the next finer level's weight is multiplied by the percent regions and subtracted from the level's weight, like so:

gapfill_scoring_weights[z_level] - z_n_pct * diff(gapfill_scoring_weights[z_level, z_level_finer])

Examples

# NOT RUN {
## setup
require(ohicore)

# gapfill
g = gapfill_georegions(data, georegions, fld_weight='w_sum')

# show result and table
head(g)
head(attr(g, 'gapfill_georegions'))
# }