United Nations Regional Groups (streamlined)

United Nations geopolitical regions (version of the georegion_labels data). In some cases, when data were not available for a country, we estimated the values using georegional averages at the smallest possible spatial scale. There are three spatial scales in these data (general to specific regions: r0_label, r1_label, r2_label).

georegions

Format

A data frame with 220 rows and 5 variables:

rgn_id

numeric region id used for Ocean Health Index (1-250)

r0

general regional groups (1 group: World)

r1

more specific regional groups (7 groups: Africa, Americas, etc.)

r2

most specific regional groups (22 groups: Caribbean, Central America, etc.)