Name to region translate (old version still used in LE function...keep until this is revised)

(old version still used in LE function...keep until this is revised)

name_to_rgn(d, fld_name = "country", flds_unique = fld_name,
  fld_value = "value", collapse_fxn = c("sum_na", "mean",
  "weighted.mean")[1], collapse_csv = NULL, collapse_flds_join = NULL,
  dir_lookup = "~/github/ohiprep/src/LookupTables",
  rgn_master.csv = file.path(dir_lookup, "eez_rgn_2013master.csv"),
  rgn_synonyms.csv = file.path(dir_lookup, "rgn_eez_v2013a_synonyms.csv"),
  add_rgn_name = F, add_rgn_type = F)

Arguments

d
dataset

fld_name
field name of the region from the dataset

flds_unique
field name for the dataset

fld_value
field with value, defaults to 'value'

collapse_fxn
function to collapse duplicate regions into one (example: China, Macau, Hong Kong)

collapse_csv
optional .csv file provided to collapse duplicate regions

collapse_flds_join
optional list of fields identified to collapse duplicate regions

dir_lookup
directory of name-to-region look up tables

rgn_master.csv
.csv file of eez-to-region combinations

rgn_synonyms.csv
.csv file of synonyms of eez-to-region combinations

add_rgn_name
T or F whether to include a column with the region name

add_rgn_type
T of F whether to include the region type (eez...)

Details

This function translates name to region id with a lookup.